Add Screen Properties.
