+++ src/screen/Profile/index.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import { Component } from 'react';
+import { View } from 'react-native';
+
+
+
+
+
+export default class ProfileScreen extends Component {
+    constructor(props) {
+        super(props);
+    }
+
+    render() {
+        return (
+            <View>
+
+            </View>
+        );
+    }
+}
